Data connectors (#23)
* draft: down-syncing

* draft: down-syncing

* draft: sqlite repos

* draft: demo repos and data

* fix: compilation

* chore: update dependencies

* chore: backend-e2e-tests setup

* test: collection-categories e2e

* test: some collections e2e

* test: some more collections e2e

* test: some more collections e2e

* test: some more collections e2e

* feat: make remote docs read only

* feat: unset remote deletes remote documents

* feat: disallow changing connectors

* test: some more collections e2e

* test: some more collections e2e

* test: some more collections e2e

* test: some more collections e2e

* test: some more collections e2e

* fix: compilation

* refactor: move test files

* fix: fix date display bug, improve demo data, run demo e2e tests in CI

* test: some more collections e2e

* chore: install all browsers

* test: some more collections e2e

* test: some more documents e2e

* test: some files e2e

* test: more e2e tests

* fix: compilation

* feat: connector authentication

* tests: more collections e2e tests

* tests: more collections e2e tests

* chore: update dependencies

* chore: update dependencies

* chore: update monaco-editor

* feat: TypeOf type util

* draft: GoogleCalendarConnector

* fix: formatting

* fix: compilation

* draft refactor: switch to OAuth2

* refactor: switch to OAuth2

* types: some improvements

* refactor: re-vibe GoogleConnector

* chore: update dependencies

* draft: set remote ui

* chore: downgrade biome

* fix: formatting

* fix: align translations

* draft: google calendar auth

* refactor: OAuth2 -> Oauth2PKCE

* tests: update down-sync tests

* feat: remote ui

* feat: sync button

* feat: remote document ui

* fix: bugs, make client_secret optional

* refactor draft: connectors

* feat: vibed GoogleContactsConnector

* refactor: no more remoteDocumentSchema

* refactor: clean up connectors

* chore: update dependencies

* fix: various fixes and refactors

* feat: auth connector button

* feat: implement with assistant

* fix: translations

* feat: implement with assistant on other fields

* fix: Strava API types

* feat: reword copy

* feat: remove unsetRemote, make remote collections remote-only, add docs

* feat: improve ai fn generation, add step for migrating converters

* draft: re-conversion instead of migration

* draft: update connectors

* draft: update demo data

* chore: update dependencies

* draft: update collections.createNewVersion

* draft: update ui

* draft: update e2e tests, fixed bug

* fix: toCamelCase

* fix: redirect after new coll version, improve ai fun impl

* fix: migration value in CreateNewCollectionVersionForm

* feat: minor prompt improvements

* feat: show required fields in form

* feat: link to remote document

* fix: error message rendering

* chore: update dependencies

* feat: improve migration perf

* refactor: more type-safe demo data

* feat: add tos to website

* test: e2e tests for website landing page

* feat: make scroll button more accessible

* chore: update ci image

* chore: upload playwright artifacts on failure

* fix: test website with astro preview

* chore: more powerful vms

* chore: ubuntu for website

* chore: smaller vms for ci

* chore: deploy website on main as well

* feat: website copy and style, minor prompt improvements

* fix: update snapshots, improve csg fn generation

* fix: css and description

* fix: tests

* chore: change macos bundle id

+++ packages/core/executing-backend/src/usecases/documents/Delete.ts
@@ -1,17 +1,19 @@
 import type {
   Backend,
   CollectionId,
+  CollectionNotFound,
   CommandConfirmationNotValid,
-  DeletedEntities,
+  ConnectorDoesNotSupportUpSyncing,
   DocumentId,
   DocumentNotFound,
   UnexpectedError,
 } from "@superego/backend";
 import type { ResultPromise } from "@superego/global-types";
-import makeDeletedEntities from "../../makers/makeDeletedEntities.js";
+import {
+  makeSuccessfulResult,
+  makeUnsuccessfulResult,
+} from "@superego/shared-utils";
 import makeResultError from "../../makers/makeResultError.js";
-import makeSuccessfulResult from "../../makers/makeSuccessfulResult.js";
-import makeUnsuccessfulResult from "../../makers/makeUnsuccessfulResult.js";
 import Usecase from "../../utils/Usecase.js";
 
 export default class DocumentsDelete extends Usecase<
@@ -21,9 +23,14 @@ export default class DocumentsDelete extends Usecase<
     collectionId: CollectionId,
     id: DocumentId,
     commandConfirmation: string,
+    allowDeletingRemoteDocument = false,
   ): ResultPromise<
-    DeletedEntities,
-    DocumentNotFound | CommandConfirmationNotValid | UnexpectedError
+    null,
+    | CollectionNotFound
+    | DocumentNotFound
+    | ConnectorDoesNotSupportUpSyncing
+    | CommandConfirmationNotValid
+    | UnexpectedError
   > {
     if (commandConfirmation !== "delete") {
       return makeUnsuccessfulResult(
@@ -34,6 +41,13 @@ export default class DocumentsDelete extends Usecase<
       );
     }
 
+    const collection = await this.repos.collection.find(collectionId);
+    if (!collection) {
+      return makeUnsuccessfulResult(
+        makeResultError("CollectionNotFound", { collectionId }),
+      );
+    }
+
     const document = await this.repos.document.find(id);
     if (!document || document.collectionId !== collectionId) {
       return makeUnsuccessfulResult(
@@ -41,17 +55,24 @@ export default class DocumentsDelete extends Usecase<
       );
     }
 
-    const deletedFileIds = await this.repos.file.deleteAllWhereDocumentIdEq(id);
-    const deletedDocumentVersionIds =
-      await this.repos.documentVersion.deleteAllWhereDocumentIdEq(id);
+    // Right now no connector supports up-syncing, so checking if the collection
+    // has a remote is sufficient. TODO: update condition once connectors
+    // support up-syncing.
+    if (collection.remote !== null && !allowDeletingRemoteDocument) {
+      return makeUnsuccessfulResult(
+        makeResultError("ConnectorDoesNotSupportUpSyncing", {
+          collectionId: collectionId,
+          connectorName: collection.remote.connector.name,
+          message:
+            "The collection has a remote, and its connector does not support up-syncing. This effectively makes the collection read-only.",
+        }),
+      );
+    }
+
+    await this.repos.file.deleteAllWhereDocumentIdEq(id);
+    await this.repos.documentVersion.deleteAllWhereDocumentIdEq(id);
     await this.repos.document.delete(id);
 
-    return makeSuccessfulResult(
-      makeDeletedEntities({
-        documents: [id],
-        documentVersion: deletedDocumentVersionIds,
-        files: deletedFileIds,
-      }),
-    );
+    return makeSuccessfulResult(null);
   }
 }
